refactor: 이벤트 리스너 분리 (김동하님께 도움받은 부분)
