+++ src/app.js
@@ -1,4 +1,5 @@
 import { getFilter, getItems } from './store.js';
 import { initRender } from './render.js';
+import './registEvent.js';
 
 initRender({ todoItems: getItems(), filter: getFilter() });
