+++ src/components/TodoViewItem.js
@@ -1,4 +1,5 @@
-const TodoViewItem = (complete, contents) => `
+const TodoViewItem = (complete, contents) =>
+    `
         <input class="toggle" type="checkbox" ${ complete && 'checked' }/>
         <label class="label">${ contents }</label>
         <button class="destroy"></button>
