+++ src/dom.js
@@ -1,0 +1,7 @@
+const $newTodoTitle = document.getElementById('new-todo-title');
+const $todoList = document.getElementById('todo-list');
+const $todoCount = document.querySelector('.todo-count strong');
+const $todoFilters = document.querySelector('.filters');
+
+
+export { $newTodoTitle, $todoList, $todoCount, $todoFilters };
